Move ArbiterProjectIdentifier into Project.h too
