+++ include/arbiter/Project.h
@@ -1,0 +1,35 @@
+#ifndef ARBITER_PROJECT_H
+#define ARBITER_PROJECT_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include <arbiter/Value.h>
+
+/**
+ * An opaque value which identifies a project participating in dependency
+ * resolution.
+ */
+typedef struct ArbiterProjectIdentifier ArbiterProjectIdentifier;
+
+/**
+ * Creates a project identifier from the given opaque data.
+ *
+ * The returned identifier must be freed with ArbiterFree().
+ */
+ArbiterProjectIdentifier *ArbiterCreateProjectIdentifier (ArbiterUserValue value);
+
+/**
+ * Returns the opaque data which was provided to ArbiterCreateProjectIdentifier().
+ *
+ * The returned pointer is only guaranteed to remain valid for the current
+ * scope.
+ */
+const void *ArbiterProjectIdentifierValue (const ArbiterProjectIdentifier *projectIdentifier);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
