+++ src/Project.h
@@ -4,10 +4,15 @@
 #error "This file must be compiled as C++."
 #endif
 
+#include <arbiter/Project.h>
+
+#include "Types.h"
+#include "Value.h"
 #include "Version.h"
 
 #include <functional>
 #include <memory>
+#include <ostream>
 #include <set>
 #include <unordered_set>
 #include <vector>
@@ -19,6 +24,30 @@ namespace Arbiter {
 class Instantiation; 
 
 } // namespace Arbiter
+
+struct ArbiterProjectIdentifier final : public Arbiter::Base
+{
+  public:
+    using Value = Arbiter::SharedUserValue<ArbiterProjectIdentifier>;
+
+    Value _value;
+
+    ArbiterProjectIdentifier ()
+    {}
+
+    explicit ArbiterProjectIdentifier (Value value)
+      : _value(std::move(value))
+    {}
+
+    std::unique_ptr<Arbiter::Base> clone () const override;
+    std::ostream &describe (std::ostream &os) const override;
+    bool operator== (const Arbiter::Base &other) const override;
+
+    bool operator< (const ArbiterProjectIdentifier &other) const
+    {
+      return _value < other._value;
+    }
+};
 
 namespace Arbiter {
